requirements.txt
backports.shutil-get-terminal-size==1.0.0
click==6.7
cycler==0.10.0
decorator==4.0.11
enum34==1.1.6
Flask==0.12.2
functools32==3.2.3.post2
ipython==5.4.1
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
matplotlib==2.0.2
MySQL-python==1.2.5
numpy==1.8.2
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.14
ptyprocess==0.5.2
PyAudio==0.2.11
PyDejavu==0.1.3
pydub==0.9.4
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
scandir==1.5
scipy==0.12.1
simplegeneric==0.8.1
six==1.10.0
subprocess32==3.2.7
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.2
